Change db to float

+++ db/migration/000002_create_batches.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "public"."batches" (
     "id" integer GENERATED ALWAYS AS IDENTITY,
     "dispatches" boolean,
-    "amount" integer,
+    "amount" float,
     "user_id" bigint,
     PRIMARY KEY ("id"),
     CONSTRAINT "user_id" FOREIGN KEY ("user_id") REFERENCES "public"."users"("id")
