+++ db/migration/000003_create_transactions.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "public"."transactions" (
     "id" integer GENERATED ALWAYS AS IDENTITY,
-    "amount" integer,
+    "amount" float,
     "user_id" bigint,
     PRIMARY KEY ("id"),
     CONSTRAINT "user_id" FOREIGN KEY ("user_id") REFERENCES "public"."users"("id")
